Fix orb when using stable-4.6 branch on GAP.dev
When running GAP.dev, GAPInfo.Version is set to "4.dev", which is higher
than any other 4.x version -- so the checks for GAP 4.7 were triggered,
even though the features those checks look for are not actually present.

Instead of looking for versions to decide which code to load, we now
look for specific features the conditionally loaded code uses.