HTValue_TreeHash_C: deal with non-int hashes, or hash = 'fail'
Since hash functions are chosen optimistically, it can happen that it
gets applied to a key value it is not prepared to handle (e.g. a hash
function might be chose which applies only to GF(2) vectors, which then
is fed a GF(4) vector.

However, this also such a key value cannot occur as a key inside the
hash. Hence, when querying the hash for the existence of such a key, we
mau simply return false. In order to determine when to do this, we need
to change hash functions to 'fail' to indicate the problem, instead of
them blindly triggering a user error.

So deal with 'fail' as a return value. And while we are it, deal with
"hash" which are neither 'fail' nor a small positive integer.

On the long run, it might be better to split this into two functions:
one which "validates" keys, and one which expects "valid" keys and
hashes them w/o further checks. This way, when e.g. resizing a hash, we
don't have to perform pointless input validation; on the other hand, we
then could validate a key without having to hash it at the same time.