Catch more bad invocations of hash functions
More hash functions now detect invalid input and return `fail`
instead of returning garbage (like hashes of random memory blocks,
or the constant 0) or even crashing.

Returning something like 0 violates the crucial invariant that
objects which compare as equal must have equal hashes. This could
happen if e.g. a hash function for compressed vectors was invoked
with input a non-compressed vector.

Also be a bit more paranoid when computing hashes of permutations,
partial permutations and transformations: just because right now
all of those have an internal rep with 2 or 4 bytes per component
does not mean this will always be the case.

This change may result in existing code to suddenly produce errors
that weren't reported before. It is almost guaranteed that this hints
at a potentially dangerous bug in the original code.